refactor: move `r_bar` and `v` parameters from Economy level to Simulation level; redefined `cap_factor` parameter as Optional[float] instead of int; update systems, validator and tests accordingly

+++ src/bamengine/events/_internal/planning.py
@@ -110,7 +110,7 @@ def firms_calc_breakeven_price(
     emp: Employer,
     lb: LoanBook,
     *,
-    cap_factor: int | None = None,
+    cap_factor: float | None = None,
 ) -> None:
     """
     Calculate breakeven price based on expected costs.
